fix(metabase): do not attempt to delete all users group on adding open database

+++ pkg/service/core/service_metabase.go
@@ -454,7 +454,7 @@ func (s *metabaseService) addAllUsersDataset(ctx context.Context, dsID uuid.UUID
 		return errs.E(op, err)
 	}
 
-	if meta.PermissionGroupID != nil {
+	if meta.PermissionGroupID != nil && *meta.PermissionGroupID == 0 {
 		err = s.metabaseAPI.DeletePermissionGroup(ctx, *meta.PermissionGroupID)
 		if err != nil {
 			return errs.E(op, err)
